feat: improve loading view in beginning of debit card module
