+++ packages/blockchain-wallet-v4-frontend/src/scenes/DebitCard/DebitCard.tsx
@@ -5,38 +5,15 @@ import { bindActionCreators, Dispatch } from 'redux'
 
 import { WalletOptionsType } from '@core/redux/walletOptions/types'
 import { RemoteDataType } from '@core/remote/types'
-import { SkeletonRectangle } from 'blockchain-info-components'
-import { Container } from 'components/Box'
 import { actions, selectors } from 'data'
 import { DebitCardType } from 'data/components/debitCard/types'
-import { useRemote } from 'hooks'
 
 import DebitCard from './DebitCard.template'
 
-const Loading = () => {
-  return (
-    <Container>
-      <SkeletonRectangle width='330px' height='270px' />
-    </Container>
-  )
-}
-
 const DebitCardContainer = (props: Props) => {
-  useEffect(() => {
-    props.debitCardActions.getCards()
-    return () => {
-      props.debitCardActions.cleanCardData()
-    }
-  }, [])
-
-  const cardsR = useRemote(selectors.components.debitCard.getCards)
-  const { data, isLoading } = cardsR
-
   if (!props.walletDebitCardEnabled) return <Redirect to='/home' />
 
-  if (isLoading) return <Loading />
-
-  return <DebitCard {...props} cards={data} />
+  return <DebitCard {...props} />
 }
 
 const mapStateToProps = (state) => ({
@@ -60,7 +37,6 @@ const connector = connect(mapStateToProps, mapDispatchToProps)
 type OwnProps = {
   alertActions: typeof actions.alerts
   cardToken: string
-  cards: Array<DebitCardType> | undefined
   debitCardActions: typeof actions.components.debitCard
   domains: { walletHelper: string }
   lockHandler: RemoteDataType<string, boolean>
